Move a basic transformation to Show a dataframe into cli/testing/transformation
